Add that if file not exist these is creted also, append conntent

+++ notes-app/app.js
@@ -4,14 +4,17 @@ const nameFile = 'notes.txt'
 const codingString = 'utf8'
 const fileNamesample = 'sample.txt'
 try {
-    fileSystem.writeFileSync(nameFile, 'My name is Nerius')
-    fileSystem.appendFileSync(nameFile, 'Perez Toirac ', codingString)
-    fileSystem.appendFileSync(nameFile, 'is begining deveop in node js, ', codingString)
-    fileSystem.appendFileSync(nameFile, ' it\svery interthing', codingString)
+    if (!fileSystem.existsSync(nameFile)) {
+        fileSystem.writeFileSync(nameFile, 'My name is Nerius')
+        fileSystem.appendFileSync(nameFile, 'Perez Toirac ', codingString)
+        fileSystem.appendFileSync(nameFile, 'is begining deveop in node js, ', codingString)
+        fileSystem.appendFileSync(nameFile, ' it\svery interthing', codingString)
+
+    }
+
 } catch (error) {
     console.log(error)
 }
-
 
 try {
     fileSystem.readFile(fileNamesample, codingString, (err, data) => {
